refactor: Piece 상속 중간 계층 추가
