+++ src/main/java/model/piece/normal/Elephant.java
@@ -1,11 +1,12 @@
-package model.piece;
+package model.piece.normal;
 
 import java.util.List;
 
 import model.Position;
 import model.Team;
+import model.piece.PieceType;
 
-public class Elephant extends Piece {
+public class Elephant extends NormalPiece {
 
     public Elephant(int x, int y, Team team) {
         super(x, y, team);
