+++ src/test/java/model/piece/PieceTest.java
@@ -10,6 +10,13 @@ import org.junit.jupiter.api.Test;
 
 import model.Team;
 import model.board.Board;
+import model.piece.iterable.Chariot;
+import model.piece.iterable.Pao;
+import model.piece.normal.Elephant;
+import model.piece.normal.Horse;
+import model.piece.normal.Palace;
+import model.piece.normal.Pawn;
+import model.piece.normal.Soldier;
 
 class PieceTest {
